pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "shogi-arena-agent"
version = "0.1.0"
description = "Runtime wrappers for connecting shogi models to external arenas."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cshogi>=0.9.7",
    "python-shogi>=1.1.1",
]

[project.optional-dependencies]
model = [
    "intelligence-representation[torch]",
]

[tool.uv.sources]
intelligence-representation = { path = "../intelligence-representation", editable = true }

[tool.setuptools.packages.find]
where = ["src"]
